# perf: reduce memory usage of background processes