--- a/frappe/defaults.py
+++ b/frappe/defaults.py
@@ -3,7 +3,6 @@
 
 import frappe
 from frappe.cache_manager import clear_defaults_cache, common_default_keys
-from frappe.desk.notifications import clear_notifications
 from frappe.query_builder import DocType
 
 # Note: DefaultValue records are identified by parent (e.g. __default, __global)