--- a/frappe/desk/form/meta.py
+++ b/frappe/desk/form/meta.py
@@ -9,7 +9,6 @@ from frappe.build import scrub_html_template
 from frappe.model.meta import Meta
 from frappe.model.utils import render_include
 from frappe.modules import get_module_path, load_doctype_module, scrub
-from frappe.translate import extract_messages_from_code, make_dict_from_messages
 from frappe.utils import get_html_format
 from frappe.utils.data import get_link_to_form
 
@@ -260,6 +259,8 @@ class FormMeta(Meta):
 				self.set("__form_grid_templates", templates)
 
 	def set_translations(self, lang):
+		from frappe.translate import extract_messages_from_code, make_dict_from_messages
+
 		self.set("__messages", frappe.get_lang_dict("doctype", self.name))
 
 		# set translations for grid templates