--- a/frappe/utils/scheduler.py
+++ b/frappe/utils/scheduler.py
@@ -15,7 +15,6 @@ from typing import NoReturn
 
 # imports - module imports
 import frappe
-from frappe.installer import update_site_config
 from frappe.utils import cint, get_datetime, get_sites, now_datetime
 from frappe.utils.background_jobs import get_jobs
 
@@ -176,6 +175,8 @@ def _get_last_modified_timestamp(doctype):
 
 @frappe.whitelist()
 def activate_scheduler():
+	from frappe.installer import update_site_config
+
 	frappe.only_for("Administrator")
 
 	if frappe.local.conf.maintenance_mode: