Because of giant `__init__.py` and huge graph of indirect imports we end up importing things that are probably never used at runtime in scheduler/background jobs. This PR removes some of these imports inside functions to defer the memory cost to when it's actually required.

- Scheduler will almost never import anything else.
- BG workers will import things in horse which is discarded so doesn't matter much. There frequently imported things staying in memory is actually a benefit. 

All units are MiB.

| process | before | after | Improvement |
| --- | --- | --- | --- |
| scheduler with one enqueue (heap from memray) | 41.1 | 25.1 | 39%|
| scheduler idle (RES from htop) | 81 | 60 | 26% | 
| BG worker (heap from memray) | 29.5 | 17.6 | 40% |



Primarily removes the following modules from the default import path:
- `requests`
- `bs4`
- `babel`
- `pytz` 
- `restrictedpython` (never required in background jobs)
- `gzip`


TODO:
- [x] analyze bg worker
- [x] add test to spot increase in idle memory usage



Tip: Use this command to quickly regenerate memory flamegraphs:

```bash
rm -f x.bin memray-flamegraph-x.html && memray run -o x.bin  ../apps/frappe/frappe/utils/bench_helper.py frappe schedule && memray flamegraph ./x.bin 
```

Brought to you by me, [memray](https://github.com/bloomberg/memray) and my Ryzen 7. 


This also reduces basic CLI overhead by 0.3 seconds just like this PR: https://github.com/frappe/frappe/pull/12447